cmd: Document the cas, refs, and init commands
Most of this is new boilerplate, but oci-image-tools.7.md is based on
the old oci-image-tool.1.md removed by fe363aa0 (*: move to
opencontainers/image-tools, 2016-09-15).  There's a lot going on in
this repo, and it's nice to have a page that outlines everything
provided by the project even though we're no longer providing a single
command.

Signed-off-by: W. Trevor King <wking@tremily.us>