Makefile: Add a pattern rule for oci-* commands
Make it easy to (re)build a single command without building all of
them.  Using main.go as the only prerequisite cuts some corners, but
prerequisites for this are hard ;).  With this commit, you can touch
main.go to force a rebuild.  If that ends up being too awkward, we can
make these .PHONY targets.

Signed-off-by: W. Trevor King <wking@tremily.us>