makefile: use full package name when building
While ./cmd/... works in the canonical Go setup, I have a very
interesting Go setup involving symlinks within my GOPATH. It turns out
that Go doesn't like this very much and the only way to solve it is to
either stop using Go's inbuilt vendoring or build everything using the
full Go import.

The sort of errors you get are that the same type cannot be used in a
function call because the function signature was imported from a
different path.

Signed-off-by: Aleksa Sarai <asarai@suse.de>