Bump runtime-spec/image-spec to v1.0.0-rc2
`v1.MediaTypeImageLayer` in latest image-spec has changed, need to be bumped up to the latest version.

This commit is only focused on bumping the vendored image-spec and runtime-spec to v1.0.0-rc2. `cobra` and `pflag` doesn't tag releases, so I'm pinning this at our currently-vendored commit. I'm not aware of any reason we can't advance this to a later pflag commit, but haven't done enough research to say for sure.

Signed-off-by: Ye Yin <eyniy@qq.com>