image: generate useful config
The previous config generation produced a config which would cause runC
to run the container using the host filesystem (with no namespaces set
up). While we should try to be as agnostic as possible, we don't want to
generate config.json payloads that are known to be bad (and useless).

Signed-off-by: Aleksa Sarai <asarai@suse.com>