image: Port to go-digest
Outsource this stuff to avoid duplication of effort.

newDescriptor was returning just the hex and createHashedBlob (its
only consumer) was fixing that (by the "Normalize the hashed digest"
comment) so it didn't have to split the hash (Hex) back off.  But that
seems confusing to me, so I've fixed newDescriptor to create a
non-busted digest which we split back apart in createHashedBlob.

Signed-off-by: W. Trevor King <wking@tremily.us>