MAINTAINERS: add Aleksa Sarai
Much of the work I've done in umoci is effectively rewriting parts of
image-tools so that they are better designed and stable. I feel very
strongly that most (if not all) of umoci should live within this
project.

As such, I feel that I'm sufficiently familiar with all of the many
implementation details necessary to correctly and sanely implement the
OCI image specification.

Signed-off-by: Aleksa Sarai <asarai@suse.de>