README: add summary of project status
We should forward users to umoci (which is now part of the OCI), because
of the many issues with oci-image-tools in the area of image
manipulation. Sadly there is no obvious home to move the validation
code, so we can just tell users to be aware that the validation code is
not being actively maintained.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>